Uporządkowanie kodu źródłowego i powtórzonych nazw
